requirements.txt
asgiref==3.4.1
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==38.0.1
defusedxml==0.7.1
Django==3.2.15
django-allauth==0.51.0
djangorestframework==3.13.1
idna==3.4
oauthlib==3.2.1
Pillow==8.4.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.12
